requirements.txt
pytest
vk_api
requests~=2.28.2
SQLAlchemy~=1.3.24
gino~=1.0.1
python-dotenv~=0.21.1